Add files via upload
Player 1 has the complete movement (can attack with shift), player 2 has the same movements but can't attack. Shelf properties just adds the basic components and layers to all shelves just add that script to a parent object where all the shelves are children of. The enemy generator only generates flies currently every 20 seconds on player 1s screen. SpatulaCol is the script attached to the spatula prefab (prefab may need to be remade to attach this script again). The spatula is shot as a projectile and can destroy enemies while ignoring the player characters and the environment (a layer may need to be added called "spatula" and "player" for the respective objects then disable collisions between the player layer and spatula layer under project settings >> 2D physics, or the spatula projectile will repel the player characters). Be sure to change the player objects layer to the "player" layer.